Chris working

+++ src/main/java/com/habibank/repo/AccountRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 import com.habibank.model.Account;
 @RepositoryRestResource
-public interface AccountRepository extends JpaRepository<Account, Integer> {
+public interface AccountRepository extends JpaRepository<Account, Long> {
 	
 }
